extractor/requirements.txt
boto3==1.42.6
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==1.1.3
langchain-google-genai==4.0.0
pymongo==4.15.4
pypdf==5.1.0
python-dotenv==1.2.2
python-multipart==0.0.20
